requirements.txt
blinker==1.9.0
blueprint==3.4.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docopt==0.6.2
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
logging==0.4.9.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pipreqs==0.4.13
python-dotenv==1.1.0
requests==2.32.3
typing==3.7.4.3
urllib3==2.3.0
Werkzeug==3.1.3
yarg==0.1.10
